+++ anxcloud/schema_lbaas.go
@@ -2,7 +2,7 @@ package anxcloud
 
 import "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-func schemaLBaaSLoadbalancer() map[string]*schema.Schema {
+func schemaLBaaSLoadBalancer() map[string]*schema.Schema {
 	return map[string]*schema.Schema{
 		"name": {
 			Type:        schema.TypeString,
@@ -12,7 +12,7 @@ func schemaLBaaSLoadbalancer() map[string]*schema.Schema {
 		"ip_address": {
 			Type:        schema.TypeString,
 			Required:    true,
-			Description: "LoadBalancer IP address.",
+			Description: "Address of the Engine management API of the LoadBalancer.",
 		},
 	}
 }