+++ CHANGELOG.md
@@ -21,6 +21,9 @@ If the change isn't user-facing but still relevant enough for a changelog entry,
 * resource/anxcloud_virtual_server: add delay after `AwaitCompletion` to handle pending changes before read (#111, @marioreggiori)
 * (internal) acceptance tests: make ProviderFactories real factories (#102, @marioreggiori)
 
+### Added
+* resource/anxcloud_lbaas_loadbalancer: add a first LBaaS resource (#107, @marioreggiori)
+
 ### Changed
 * resource/anxcloud_virtual_server: increase delete timeout (#112, @marioreggiori)
 * (internal) acceptance tests: configured to run parallel (#102, @marioreggiori)